backend/mitgliederverwaltung/requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
backcall==0.1.0
cffi==1.11.5
cryptography==2.4.2
cssselect2==0.2.1
decorator==4.3.0
Django==2.0.3
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-rest-knox==3.4.0
django-webpack-loader==0.6.0
djangorestframework==3.9.0
idna==2.8
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.3.3
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.15.4
pandas==0.22.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.6.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
pylint==2.1.1
pyOpenSSL==18.0.0
PyPDF2==1.26.0
python-dateutil==2.7.5
pytz==2018.7
qrcode==6.1
reportlab==3.5.12
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
tinycss2==0.6.1
traitlets==4.3.2
typed-ast==1.1.0
typing==3.6.6
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
